Add url params setting and loading. Make main async
